re-introduce derive tests outside of the non-derive crate
Some derive tests do not yet work due to macro issues.
It seems that the macros parse tuple-structs incorrectly.
